requirements.txt
django>=4.0.0,<4.1.0
django-cors-headers==3.12.0
djangorestframework==3.13.1
drf-access-policy==1.1.1
drf-nested-routers==0.93.4
djangorestframework-simplejwt==5.2.0
idna==3.3
isort==5.10.1
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.5.2
PyJWT==2.4.0
pytz==2022.1
PyYAML==6.0
sqlparse==0.4.2
tomli==2.0.1
urllib3==1.26.9